+++ panther/cli/main.py
@@ -1,5 +1,4 @@
 import os
-import platform
 import sys
 
 from rich import print as rprint
@@ -8,14 +7,36 @@ from panther import version as panther_version
 from panther.cli.create_command import create
 from panther.cli.monitor_command import monitor
 from panther.cli.run_command import run
-from panther.cli.utils import clean_args, cli_error, help_message
+from panther.cli.utils import clean_args, cli_error, cli_info, cli_warning, help_message
 
 
-def shell() -> None:
-    if platform.system().lower() == 'windows':
+def shell(args: list) -> None:
+    if len(args) == 0:
+        cli_info('You may want to use "bpython" or "ipython" for better interactive shell')
         os.system('python')
-    else:
-        os.system('bpython')
+    elif len(args) != 1:
+        return cli_error('Too Many Arguments.')
+    shell_type = args[0].lower()
+    if shell_type not in ['ipython', 'bpython']:
+        return cli_error(f'"{shell_type}" Is Not Supported.')
+
+    # Bpython
+    if shell_type == 'bpython':
+        try:
+            import bpython
+            os.system('bpython')
+        except ImportError as e:
+            cli_warning(e, 'Hint: "pip install ipython"')
+            os.system('python')
+
+    # Ipython
+    elif shell_type == 'ipython':
+        try:
+            import IPython
+            os.system('ipython')
+        except ImportError as e:
+            cli_warning(e, 'Hint: "pip install bpython"')
+            os.system('python')
 
 
 def version() -> None:
@@ -37,7 +58,7 @@ def start() -> None:
             case 'run':
                 run(args)
             case 'shell':
-                shell()
+                shell(sys.argv[2:])
             case 'monitor':
                 monitor()
             case 'version':